Refactor Results and implement readStream
This commit refactors the softDelete operation to complete exceptionally
in the case of errors (primarily `WrongExpectedVersionException`).

It also adds an implementation of readStream and the associated types
and tests. In order to test the wire protocol to public API type
mapping, we add a JSON file containing the serialized events for each
test, and a "Test*" class heirarchy compatible with the Jackson JSON
parser with equality defined against the types in the public API.