Introduce Timeouts & TimeoutsBuilder
This commit introduces a new class, `Timeouts` which must be passed to
instances of `StreamsClient`. A static final `DEFAULT` instance can be
passed in place of defining custom timeouts. A TimeoutsBuilder object
provides an extensible way to build this value as new operations which
require timeouts get added, without breaking forwards compatibility.