Add subscribeToStream support
This commit adds support for the `subscribeToStream` operation to the
`StreamsClient`. This returns a `CompletableFuture<Subscription>` which
completes when the subscription is confirmed by the server.

Events, errors and cancellation notices are delivered to the supplied
instance of `SubscriptionListener` (which itself is abstract and
contains default implementations of each handler).

Tests are added for a variety of conditions:

- user-initiated cancellation of a subscription after delivery of some
  events
- user-initiated cancellation of a subscription after all events in the
  stream have been delivered
- user-initiated cancellation of a subscription after all events in the
  stream have been delivered, a new event appended to the stream and
  delivery to the subscription verified.

Many of the same types and patterns can be used for subscriptions to
`$all`.