Add subscribeToAll operation to StreamsClient
This commit adds the subscribeToAll operation to the `StreamsClient`
class. In order to correctly support the protocol, a
`SubscriptionFilter` type is introduced which allows specification of a
`Checkpointer` for recording checkpoints where events do not match the
filter.

As part of this, we rework adding the filter to an outbound `ReadReq`
making better use of object oriented programming rather than static
methods.

Testing is currently somewhat rudimentary but was sufficient to confirm
basic operation as well as find some bugs:

- we test 1000 events before cancelling the read from `$all`
- we perform a filtered read on a stream, comparing the positions and
  event types with the expected values from the known dataset.