Add SubscriptionFilterBuilder class
Building `SubscriptionFilter` instances is somewhat involved - it's
necessary to construct either a `StreamFilter` or `EventTypeFilter` with
the appropriate Prefix or Regular expressions, and configure checkpoints
and windowing.

This commit adds a `SubscriptionFilterBuilder` to reduce the amount of
boilerplate needed for this.

For example, this construction:

```
SubscriptionFilter filter = new SubscriptionFilter(
  new EventTypeFilter(Optional.empty(),
    new RegularFilterExpression(Pattern.compile("^eventType-194$"))));
```

Becomes this builder instantiation:

```
SubscriptionFilter filter = SubscriptionFilter.newBuilder()
  .withEventTypeRegularExpression("^eventType-194$")
  .build();
```

The difference is more acute when a `Checkpointer` and optional
non-default interval is specified.

The previous mechanism is still available (and used internally by the
builder) for if unusual options or a more manual construction system is
needed.